src/requirements.txt
requests==2.32.3
feedparser==6.0.10
PyYAML==6.0.1
pycti==6.6.17